Pin electron version to 25.1.1
Normally electron-builder will look inside `node_modules` to detect the version of electron installed, but it's not likely to look for hoisted one, and falls back to search from `dependencies` entry in package.json.
This change is a workaround for the following error:

Cannot compute electron version from installed node modules - none of the possible electron modules are installed and version ("^25.1.1") is not fixed in project.

See https://github.com/electron-userland/electron-builder/issues/3984#issuecomment-504968246